python_cncagent/requirements.txt
PyMuPDF==1.23.8
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
spacy==3.6.1
torch==2.1.0
transformers==4.34.0
flask==2.3.3
flask-cors==4.0.0